refactor: main electron file
